apps/backend/app/requirements.txt
# --- Web ---
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.21

# --- Settings / config ---
pydantic==2.10.4
pydantic-settings==2.6.1
python-dotenv==1.0.1

# --- DB core ---
SQLAlchemy==2.0.36
alembic==1.14.0

# --- PostgreSQL driver ---
asyncpg==0.30.0
# --- PostGIS / geometry ---
GeoAlchemy2==0.16.0
shapely==2.0.6

# --- Useful utils (optional but common) ---
httpx==0.28.1
orjson==3.10.12

# --- Auth ---
python-jose[cryptography]==3.5.0
